Create D15090001
Reset the values to null for "Show Total Cash Held" and "Press to Collect Cash" when the maintainer successfully pushes "Press Here When Finished". The simulation currently does not clear the values on the completion of a Maintenance session. Upon a subsequent re-login to the Maintenance Panel in the same session, the cash previously held and already collected, persists in display. This sometimes causes confusion to Maintainer. Need to be reset.